Avoid duplicate files for multiple replicas
When a file has multiple replicas in iRODS, rclOpenCollection would
return multiple entries for the file.

Do the same fix as what Jargon does in
CollectionAndDataObjectListAndSearchAOImpl.buildDataObjectListingWithAccessInfoFromResultSet:
remember the last name seen, and only process the entry as a new file if
it is different from the last name seen.

(Note: it is OK to keep a copy of the pointer while inside
iRODS_l_stat_dir - the points are returned to globus and are the
responsiblity of globus to free only after this function returns).