Include "." and ".." in directory listings.
So far, a "." entry was included only when the directory listing would
be otherwise empty (and for the root folder where the self-refering
entry is already included in the iRODS listing).

Instead of adding it only for empty listing, always start the listing
with a "." and ".." - but not for the root directory which would already
get the "." anyway - and does not need a "..".

Note that these two special entries would not have their timestamp set -
this would require two additional look up calls to the server and that's
not worth the performance and code complexity penalty.