homeDir: use parsed username without zone
Use parsed username (without zone) when constructing the home directory
with homeDirPattern.

Otherwise, when the user is authenticated with a full username
(including zone), finished_info.info.session.home_dir would be set to a
directory that does not exist (/<zone>/home/<user>#<zone>/).

Fix this by setting it to /<zone>/home/<user>/ (or whatever
homeDirPattern specifies, substituting first %s with zone and second %s
with username with the zone specified stripped off).