New feature: server connects as rods, acts as user
Add an alternative way of connecting to iRODS, with a variant of the
rcConnect call where the client authenticates as a rodsAdmin user (rods)
and specifies to act on behalf of a particular user.
(ProxyUserName="rods", ClientUserName=iRODS_handle->user)

Add a new environment variable $irodsConnectAsAdmin controlling this
feature.

If this variable is absent, authentication is unchanged, with the
original rcConnect call.

If this variable is set, authentication is done via _rcConnect, passing
the username from getRodsEnv as ProxyUserName (authenticating against
this account with the server certificate) and passing iRODS_handle->user
as ClientUserName.

This eliminates the need to add the server DN to each user account - it
is enough to just once add the server DN to the rods account.

Feature works only when the server certificate has access to the rods
account (or whatever rodsAdmin account is specified in the iRODS
environment).

Update documentation accordingly.

NOTE: Feature works only with iRODS 3.3 and above (support for proxyUser with
GSI only added in 3.3; breaks with iRODS 3.2 and earlier).